+++ NEWS.md
@@ -3,7 +3,12 @@ NEWS-KDHEtools
 
 <!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
 
-    #> Last Update: 2023-08-09 15:42:17.755668
+    #> Last Update: 2023-08-10 07:00:35.357436
+
+# KDHEtools v0.1.2.9013 (2023-08-09)
+
+- docs: Remove unused packages from DESCRIPTION and global.R
+  - Predictors, Blocktt/KDHEtools#7
 
 # KDHEtools v0.1.2.9012 (2023-08-09)